% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchr-package.R
\docType{package}
\name{dispatchr-package}
\alias{dispatchr}
\alias{dispatchr-package}
\title{dispatchr: Lightweight Outbound Messaging}
\description{
Small helpers for sending outbound messages through Gmail SMTP, X, and
Telegram.
}
\author{
\strong{Maintainer}: Oliver Zhou \email{oliver.yxzhou@gmail.com}

Authors:
\itemize{
  \item Oliver Zhou \email{oliver.yxzhou@gmail.com}
}

}
\keyword{internal}
